add status to engagement summary to the card
