+++ src/components/engagement_data_cards/engagement_summary_card/engagement_summary_card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DataCard } from '../data_card';
-import { Engagement } from '../../../schemas/engagement_schema';
+import { Engagement, getEngagementStatus } from '../../../schemas/engagement_schema';
 import { Grid, GridItem } from '@patternfly/react-core';
 import { TitledDataPoint } from '../../titled_data_point/titled_data_point';
 import { format as formatDate } from 'date-fns';
@@ -8,6 +8,7 @@ import { EngagementSummaryEditModal } from '../../engagement_edit_modals/engagem
 import { useModalVisibility } from '../../../context/edit_modal_visibility_context/edit_modal_visibility_hook';
 import { EditButton } from '../../data_card_edit_button/data_card_edit_button';
 import { RequiredFieldsWarning } from '../../required_fields_warning/required_fields_warning';
+import { EngagementStatusText } from "../../../routes/engagement_list/engagement_status_text";
 
 export interface EngagementSummaryCardProps {
   engagement: Engagement;
@@ -30,6 +31,8 @@ export function EngagementSummaryCard({
     'start_date',
     'project_name',
   ];
+  const status = getEngagementStatus(engagement);
+
   return (
     <>
       <EngagementSummaryEditModal
@@ -58,38 +61,48 @@ export function EngagementSummaryCard({
         title="Engagement Summary"
       >
         <Grid hasGutter>
-          <GridItem span={4}>
+          <GridItem style={{marginBottom: '1rem'}}>
+            <EngagementStatusText status={status} />
+          </GridItem>
+          <GridItem span={3}>
             <TitledDataPoint title="Company">
               {engagement?.customer_name}
             </TitledDataPoint>
           </GridItem>
-          <GridItem span={4}>
+          <GridItem span={3}>
             <TitledDataPoint title="Project">
               {engagement?.project_name}
             </TitledDataPoint>
           </GridItem>
-          <GridItem span={4}>
+          <GridItem span={3}>
             <TitledDataPoint title="Location">
               {engagement.location}
             </TitledDataPoint>
           </GridItem>
-          <GridItem span={4}>
+          <GridItem span={3}>
+            <TitledDataPoint title="Description">
+              {engagement?.description}
+            </TitledDataPoint>
+          </GridItem>
+          <GridItem span={3}>
             <TitledDataPoint title="Start Date">
               {engagement?.start_date
                 ? formatDate(engagement?.start_date, 'MMM dd, yyyy')
                 : null}
             </TitledDataPoint>
           </GridItem>
-          <GridItem span={4}>
+          <GridItem span={3}>
             <TitledDataPoint title="End Date">
               {engagement?.end_date
                 ? formatDate(engagement?.end_date, 'MMM dd, yyyy')
                 : null}
             </TitledDataPoint>
           </GridItem>
-          <GridItem span={4}>
-            <TitledDataPoint title="Description">
-              {engagement?.description}
+          <GridItem span={3}>
+            <TitledDataPoint title="Created By">
+              {!!engagement?.creation_details?.created_by_user
+                ? engagement?.creation_details?.created_by_user
+                : ''}
             </TitledDataPoint>
           </GridItem>
         </Grid>
