+++ src/routes/engagement_list/engagement_status_text.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import {EngagementStatus} from "../../schemas/engagement_schema";
+
+export const EngagementStatusText = ({ status }: { status: EngagementStatus }) => {
+  const getEngagementStatusText = () => {
+    if (status === EngagementStatus.active) {
+      return 'Active';
+    } else if (status === EngagementStatus.upcoming) {
+      return 'Upcoming';
+    } else if (status === EngagementStatus.past) {
+      return 'Past';
+    }
+    return '';
+  };
+  const getStatusColor = (status?: EngagementStatus) => {
+    switch (status) {
+      case EngagementStatus.upcoming: {
+        return '#EC7A08';
+      }
+      case EngagementStatus.active: {
+        return 'green';
+      }
+      default: {
+        return '#B8BBBE';
+      }
+    }
+  };
+  return (
+    <b style={{ color: getStatusColor(status) }}>
+      {getEngagementStatusText().toUpperCase()}
+    </b>
+  );
+};
